add import time check script
