+++ .evergreen/run-import-time-test.sh
@@ -1,0 +1,39 @@
+#!/bin/bash -ex
+
+set -o errexit  # Exit the script with error if any of the commands fail
+set -x
+
+. .evergreen/utils.sh
+
+if [ -z "$PYTHON_BINARY" ]; then
+    PYTHON_BINARY=$(find_python3)
+fi
+
+# Skip the report of it isn't a PR run.
+if [ "$BASE_SHA" == "$HEAD_SHA" ]; then
+    echo "Skipping API Report"
+    exit 0
+fi
+
+function get_import_time() {
+    createvirtualenv "$PYTHON_BINARY" import-venv
+    python -m pip install -q ".[aws,encryption,gssapi,ocsp,snappy,zstd]"
+    # Import once to cache modules
+    python -c "import pymongo"
+    local log_file="pymongo-$1.log"
+    python -X importtime -c "import pymongo" 2> $log_file
+    local last_line=$(echo $(tail -n 1 $log_file) | cut -d " " -f 5)
+    rm -rf import-venv
+    echo $last_line
+}
+
+import_time_curr=$(generate_import_log $HEAD_SHA)
+git checkout $BASE_SHA
+import_time_prev=$(generate_import_log $BASE_SHA)
+
+# Check if we got 20% or more slower
+let ratio=($import_time_curr-$import_time_prev / $import_time_prev
+if [ $ratio -gt 0.2 ]; then
+    echo "Import got $ratio percent slower"
+    exit 1
+fi
